+++ src/admin/hooks/hooksUI/index.ts
@@ -1,4 +1,8 @@
 export { useCourseView } from "./Course/useCourseView";
 export { useCreateCourseView } from "./Course/useCreateCourseView";
+
 export { useCreateYearView } from "./Year/useCreateYearView";
 export { useListYear } from "./Year/useListYear";
+
+export { useListTeacherView } from "./Teacher/useListTeacherView";
+export { useCreateTeacherView } from "./Teacher/useCreateTeacherView";
